TOR-19: address review — extract daily_pick_service, real refresh, hardening
Review findings addressed:
1. refresh regeneration now injects the rejected pick into the prompt
   ("REJECTED PICK: ... generate a meaningfully DIFFERENT workout")
2. get_daily_recommendation gains a refresh param wired to the shared
   generation path, so "give me something different" actually regenerates
   (and the dashboard shows the new pick); prompts updated to say so
3. explicit `is`-checks so JSON null for optionals keeps defaults
4. recent-picks dates derived from local_date (the cache key) instead of a
   second, differently-sourced timezone resolution
5. generated_at guarded with isinstance, matching the endpoint
6+8. generation logic, per-user locks, and local-date resolution moved out of
   the API route into app/services/daily_pick_service.py with a single public
   get_or_generate_today_pick(); local-date now delegates to the shared
   date_utils.get_user_today; the skill no longer imports api privates;
   route module re-exports helpers for back-compat
7. rest gate tightened: 3+ consecutive completed days must END YESTERDAY
9. get_recent returns None on fetch error; the chat context omits the
   Today's Pick block instead of falsely asserting no pick exists
10. ellipsis only when exercise names were actually dropped; cap exhaustion
   before a later block is announced explicitly

331 tests pass (11 new); skill re-verified read-only against live data.
