+++ ai-coach-service/app/core/agents/prompts.py
@@ -83,7 +83,7 @@ WHEN USER ASKS ABOUT EXERCISES BY MUSCLE GROUP (e.g., "what core exercises do I 
 - `adjust_plan`: Change a live plan's volume/frequency or add a deload mid-cycle. Previews + re-validates; big volume jumps need override.
 
 **Daily Suggestion ("Today's Pick")**:
-- `get_daily_recommendation`: the daily AI-suggested workout shown on the user's Dashboard / Train Now page. Fetches its full exercises/sets (and generates one if today's doesn't exist yet — the dashboard will show the same one). This is THE answer to "what should I do today?" when the calendar is empty. NEVER invent a different workout for today without acknowledging the existing pick; if the user wants something different, offer to refresh it so chat and dashboard stay consistent.
+- `get_daily_recommendation`: the daily AI-suggested workout shown on the user's Dashboard / Train Now page. Fetches its full exercises/sets (and generates one if today's doesn't exist yet — the dashboard will show the same one). This is THE answer to "what should I do today?" when the calendar is empty. NEVER invent a different workout for today without acknowledging the existing pick; if the user rejects it or wants something different, call it again with `refresh=true` to regenerate — the dashboard updates to the new pick too, so chat and dashboard stay consistent.
 
 **Web Search & Research** (External resources):
